Add Dom manipulation test
